aio: fix a compiler warning on 32 bit architectures in aio_renameat()
The kbuild test robot reported a compiler warning:

   fs/aio.c: In function 'aio_renameat':
>> fs/aio.c:1995:14: warning: cast to pointer from integer of different
size
[-Wint-to-pointer-cast]
     user_info = (const void * __user)user_iocb->aio_buf;

A cast was missed in aio_renameat() that leads to a warning about a
pointer cast from a different size integer.  Add a cast to suppress this
issue on 32 bit architectures.

+++ fs/aio.c
@@ -1992,7 +1992,7 @@ long aio_renameat(struct aio_kiocb *iocb, struct iocb *user_iocb)
 	if (user_iocb->aio_offset)
 		return -EINVAL;
 
-	user_info = (const void * __user)user_iocb->aio_buf;
+	user_info = (const void * __user)(long)user_iocb->aio_buf;
 	if (unlikely(!access_ok(VERIFY_READ, user_info,
 				sizeof(struct renameat_info))))
 		return -EFAULT;
